Updated property inspector to refresh (using react keys) on model create / delete; small logic and renaming changes
