+++ client/src/components/PropertyInspectorComponent/PropertyInspectorComponent.scss
@@ -1,218 +1,26 @@
-// Copyright (c) Microsoft Corporation.
-// Licensed under the MIT license.
-
 @import "../../theme/theme.scss";
 
-:root {
-  --json-editor-empty-value: "Unset";
-}
-
-.pi-gridWrapper {
-  height: inherit;
-  max-height: 100%;
-  .ms-Button-label {
-    font-weight: 200 !important;
-  }
-}
-
 .pi-patch-modal {
-  h2 {
-    font-weight: 400;
-    margin-bottom: 10px;
-    margin-top: 0;
-  }
-  .modal-button {
-    width: 100px;
-    margin-right: 10px;
-    color: $textColorLight1;
-    font-weight: 600;
-    border: none;
-    &.close-button {
-      float: right;
-      background: $buttonColor;
+    h2 {
+        font-weight: 400;
+        margin-bottom: 10px;
+        margin-top: 0;
     }
-  }
+    .modal-button {
+        width: 100px;
+        margin-right: 10px;
+        color: $textColorLight1;
+        font-weight: 600;
+        border: none;
+        &.close-button {
+        float: right;
+        background: $buttonColor;
+        }
+    }
 }
 
-.pi-grid {
-  display: grid;
-  grid-template-rows: auto auto 1fr;
-  height: 100%;
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 0;
-  bottom: 0;
-  overflow: auto;
-
-  .pi-title {
-    font-size: 10px;
-  }
-}
-
-.pi-toolbarButtons {
-  height: 16px;
-  font-size: 8px;
-  font-weight: 200;
-  &:hover, &:focus {
-    background: transparent;
-  }
-}
-
-.pi-editor {
-  height: 100%;
-  overflow: auto;
-  padding-left: 16px;
-  font-weight: 200;
-}
-
-.jsoneditor {
-  height: 100%;
-  max-height: 100%;
-  border: 0px;
-  margin-top: 26px;
-}
-
-.pi-filter {
-  font-size: 12px;
-  font-weight: 200;
-  padding: 0 15px;
-
-  .ms-TextField-field {
-    font-family: $fontFamilyMain;
-    font-size: 14px;
-    background-color: #353535;
-    color: $textColorLight1;
-    border-width: 0px;
-    flex: 1 1;
-  }
-}
-
-.pi-command-bar {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding-left: 16px;
-  color: #ccc;
-  .pi-title {
-    text-transform: uppercase;
-    color: var(--bg);
-  }
-}
-
-.jsoneditor-expandable:nth-child(2) {
-  width: 0px;
-}
-
-.jsoneditor-expandable:first-child {
-  width: 0px;
-}
-
-.jsoneditor-tree .jsoneditor-value, .jsoneditor-tree .jsoneditor-field {
-  user-select: all;
-}
-
-.jsoneditor-tree > colgroup > col:first-child {
-  width: 0px;
-}
-
-.jsoneditor-tree > colgroup > col:nth-child(2) {
-  width: 0px;
-}
-
-.jsoneditor-contextmenu-button {
-  display: none;
-}
-
-.jsoneditor-frame > input {
-  font-family: $fontFamilyMain !important;
-  font-weight: 600 !important;
-  font-size: 12px !important;
-  background-color: $inputColor;
-  color: $textColorLight2;
-  border-width: 0px !important;
-  flex: 1;
-}
-
-.jsoneditor-results {
-  font-family: $fontFamilyMain;
-  font-weight: 200;
-  font-size: 10px;
-  color: $textColorLight1;
-}
-
-.jsoneditor-navigation-bar {
-  display: none;
-}
-
-.jsoneditor-treepath .jsoneditor-treepath-element {
-  font-family: $fontFamilyMain;
-  font-weight: 400;
-  font-size: 12px;
-}
-
-.jsoneditor-field {
-  color: $textColorLight1 !important;
-}
-
-div.jsoneditor-tree button.jsoneditor-button:focus {
-  background-color: $backgroundColor2;
-  outline: $outlineColor4 solid 0px;
-}
-
-div.jsoneditor-value.jsoneditor-string {
-  color: $editorStringColor;
-}
-
-div.jsoneditor-value.jsoneditor-number {
-  color: $editorNumberColor;
-}
-
-div.jsoneditor-field, div.jsoneditor-value, div.jsoneditor td, div.jsoneditor th, div.jsoneditor textarea, pre.jsoneditor-preview, .jsoneditor-schema-error, .jsoneditor-popover {
-  font-family: $fontFamilyMain;
-  color: $textColorLight1;
-}
-
-div.jsoneditor-readonly {
-  color: $textColorLight1;
-  font-weight: 400;
-}
-
-div.jsoneditor-object {
-  font-weight: 600;
-}
-
-div.jsoneditor-value {
-  font-family: $fontFamilyConsole;
-}
-
-.jsoneditor-telemetry {
-  .jsoneditor-value {
-    color: $editorStringColor !important;
-  }
-}
-
-.jsoneditor-expandable {
-  .jsoneditor-field {
-    font-weight: 400;
-  }
-}
-
-div.jsoneditor-field[contenteditable="true"]:focus, div.jsoneditor-field[contenteditable="true"]:hover, div.jsoneditor-value[contenteditable="true"]:focus, div.jsoneditor-value[contenteditable="true"]:hover, div.jsoneditor-field.jsoneditor-highlight, div.jsoneditor-value.jsoneditor-highlight {
-  background-color: $inputColor;
-  border: 1px solid $inputOutlineColor;
-  border-radius: 0;
-  min-height: 12px;
-}
-div.jsoneditor-value.jsoneditor-number, div.jsoneditor-value.jsoneditor-string {
-  color: $textColorLight1;
-}
-
-.jsoneditor-menu > button:hover, .jsoneditor-menu > .jsoneditor-modes > button:hover {
-  background-color: transparent;
-  border: 1px solid transparent;
-}
-
-div.jsoneditor-value.jsoneditor-empty::after {
-  content: var(--json-editor-empty-value);
+.property-inspector-container{
+    height: calc(100% - 8px);
+    padding: 8px 16px 20px 16px;
+    overflow: hidden;
 }
