--- client/src/components/PropertyInspectorComponent/PropertyInspectorComponentV2.scss
@@ -1,26 +1,0 @@
-@import "../../theme/theme.scss";
-
-.pi-patch-modal {
-    h2 {
-        font-weight: 400;
-        margin-bottom: 10px;
-        margin-top: 0;
-    }
-    .modal-button {
-        width: 100px;
-        margin-right: 10px;
-        color: $textColorLight1;
-        font-weight: 600;
-        border: none;
-        &.close-button {
-        float: right;
-        background: $buttonColor;
-        }
-    }
-}
-
-.property-inspector-container{
-    height: calc(100% - 8px);
-    padding: 8px 16px 20px 16px;
-    overflow: hidden;
-}
